+++ tests/test_basePDF.py
@@ -6,6 +6,7 @@ import numpy as np
 
 import zfit.core.basepdf
 from zfit.core.limits import Range
+import zfit.pdfs.dist_tfp
 from zfit.pdfs.dist_tfp import Normal
 from zfit.pdfs.basic import Gauss
 from zfit.core.parameter import FitParameter
@@ -32,7 +33,7 @@ def true_gaussian_func(x):
 mu2 = FitParameter("mu", mu_true, mu_true - 2., mu_true + 7.)
 sigma2 = FitParameter("sigma", sigma_true, sigma_true - 10., sigma_true + 5.)
 tf_gauss1 = tf.distributions.Normal(loc=mu2, scale=sigma2, name="tf_gauss1")
-wrapped_gauss = zfit.core.basepdf.WrapDistribution(tf_gauss1)
+wrapped_gauss = zfit.pdfs.dist_tfp.WrapDistribution(tf_gauss1)
 
 test_gauss1 = TestGaussian(name="test_gauss1")
 wrapped_normal1 = Normal(loc=mu2, scale=sigma2, name='wrapped_normal1')
