+++ zfit/core/basepdf.py
@@ -1098,50 +1098,3 @@ class BasePDF(pep487.PEP487Object):
             raise TypeError("Cannot compare objects of type {} and {}".format(type(self), type(other)))
         params_equal = set(other.parameters) == set(self.parameters)
         return params_equal
-
-
-class WrapDistribution(BasePDF):
-    """Baseclass to wrap tensorflow-probability distributions automatically.
-
-    """
-
-    def __init__(self, distribution, name=None, **kwargs):
-        # Check if subclass of distribution?
-        name = name or distribution.name
-        super(WrapDistribution, self).__init__(distribution=distribution, name=name, **kwargs)
-        # self.tf_distribution = self.parameters['distribution']
-        self.tf_distribution = distribution
-
-    def _unnormalized_prob(self, x):
-        return self.tf_distribution.prob(value=x, name="_unnormalized_prob")  # TODO name
-
-    # TODO: register integral
-    @no_norm_range
-    def _analytic_integrate(self, limits, norm_range):  # TODO deal with norm_range
-        lower, upper = limits.get_boundaries()  # TODO: limits
-        upper = tf.cast(upper, dtype=tf.float64)
-        lower = tf.cast(lower, dtype=tf.float64)
-        integral = self.tf_distribution.cdf(upper) - self.tf_distribution.cdf(lower)
-        return integral
-
-
-# TODO: remove below, play around while developing
-if __name__ == "__main":
-    import zfit
-
-    mu_true = 1.4
-    sigma_true = 1.8
-
-
-    class TestGaussian(zfit.core.basepdf.BasePDF):
-        def _unnormalized_prob(self, x):
-            return tf.exp(-(x - mu_true) ** 2 / sigma_true ** 2)  # non-normalized gaussian
-
-
-    dist1 = TestGaussian()
-    tf_gauss1 = tf.distributions.Normal(loc=mu_true, scale=sigma_true)
-    wrapped = WrapDistribution(tf_gauss1)
-
-    with tf.Session() as sess:
-        res = sess.run(dist1.event_shape_tensor())
-        print(res)
